# geral/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "SH3 API Laravel",
        "description": "SH3 - API de serviços independentes",
        "contact": {
            "name": "SH3 Suporte",
            "email": "suporte@sh3.com.br"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:{porta}/api/",
            "description": "Servidor de desenvolvimento",
            "variables": {
                "porta": {
                    "enum": [
                        "8000",
                        "8001"
                    ],
                    "default": "8000"
                }
            }
        }
    ],
    "paths": {
...
        "/geral/endereco/": {
            "get": {
                "tags": [
                    "endereco"
                ],
                "summary": "Listar enderecos",
                "operationId": "listarEnderecos",
                "responses": {
                    "200": {
                        "description": "Lista de enderecos",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CadastroEndereco"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            },
            "post": {
                "tags": [
                    "endereco"
                ],
                "summary": "Inserir endereco",
                "operationId": "inserirEnderecos",
                "requestBody": {
                    "description": "Informar dados do endereco",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CadastroEndereco"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereco criado com sucesso"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
        "/geral/endereco/{id}": {
            "put": {
                "tags": [
                    "endereco"
                ],
                "summary": "Atualizar endereco",
                "operationId": "updateEnderecos",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Id do endereço",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Informar dados do endereco",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CadastroEndereco"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Endereco editado com sucesso"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
        "/geral/municipio/": {
            "get": {
                "tags": [
                    "municipio"
                ],
                "summary": "Listar municipios",
                "operationId": "listarMunicipios",
                "responses": {
                    "200": {
                        "description": "Lista de municipios",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/CadastroMunicipio"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            },
            "post": {
                "tags": [
                    "municipio"
                ],
                "summary": "Inserir municipio",
                "operationId": "inserirMunicipios",
                "requestBody": {
                    "description": "Informar dados do municipio",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CadastroMunicipio"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Municipio criado com sucesso"
                    },
                    "401": {
                        "$ref": "#/components/responses/UnauthorizedError"
                    },
                    "500": {
                        "$ref": "#/components/responses/UnexpectedError"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "CadastroEndereco": {
                "title": "ENDERECO",
                "required": [
                    "INT_MUN"
                ],
                "properties": {
                    "INT_END": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "INT_MUN": {
                        "type": "integer",
                        "example": "1"
                    }
                },
                "type": "object"
            },
...
            "CadastroMunicipio": {
                "title": "MUNICIPIO",
                "required": [
                    "NM_MUN",
                    "INT_IBGE"
                ],
                "properties": {
                    "INT_MUN": {
                        "type": "integer",
                        "readOnly": true,
                        "example": 1
                    },
                    "NM_MUN": {
                        "type": "string",
                        "maxLength": 255,
                        "example": "Teste"
                    },
                    "INT_IBGE": {
                        "type": "integer",
                        "example": "123456"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
            "UnauthorizedError": {
                "description": "Token de acesso não encontrado ou inválido",
                "content": {
                    "application/json": {
                        "example": {
                            "status": "error",
                            "message": "Token not found."
                        }
                    }
                }
            },
            "UnexpectedError": {
                "description": "Erro inesperado",
                "content": {
                    "application/json": {
                        "example": {
                            "status": "error",
                            "message": "An exception occurred while executing a query: [...]"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "JWT": {
                "type": "http",
                "name": "JWT",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}